+++ library/src/transformations/toMaxValue/toMaxValue.ts
@@ -1,4 +1,5 @@
 import type { PipeResult } from '../../types.ts';
+import { getOutput } from '../../utils/index.ts';
 
 /**
  * Creates a transformation function that sets a string, number or date to a
@@ -12,7 +13,6 @@ export function toMaxValue<
   TInput extends string | number | bigint | Date,
   TRequirement extends TInput
 >(requirement: TRequirement) {
-  return (input: TInput): PipeResult<TInput> => ({
-    output: input > requirement ? requirement : input,
-  });
+  return (input: TInput): PipeResult<TInput> =>
+    getOutput(input > requirement ? requirement : input);
 }